+++ src/Swashbuckle.AspNetCore.Filters/Examples/ExamplesConverter.cs
@@ -1,34 +1,45 @@
+using Microsoft.Net.Http.Headers;
+using Microsoft.OpenApi.Any;
+using Microsoft.OpenApi.Models;
+using Swashbuckle.AspNetCore.Filters.Extensions;
 using System;
 using System.Collections.Generic;
 using System.Linq;
-using Microsoft.OpenApi.Any;
-using Microsoft.OpenApi.Models;
-using Newtonsoft.Json;
-using Swashbuckle.AspNetCore.Filters.Extensions;
 
 namespace Swashbuckle.AspNetCore.Filters
 {
     internal class ExamplesConverter
     {
-        private readonly JsonFormatter jsonFormatter;
+        private static readonly MediaTypeHeaderValue ApplicationXml = MediaTypeHeaderValue.Parse("application/xml; charset=utf-8");
+        private static readonly MediaTypeHeaderValue ApplicationJson = MediaTypeHeaderValue.Parse("application/json; charset=utf-8");
+
         private readonly MvcOutputFormatter mvcOutputFormatter;
-        private readonly JsonSerializerSettings serializerSettings;
 
-        public ExamplesConverter(JsonFormatter jsonFormatter, MvcOutputFormatter mvcOutputFormatter, JsonSerializerSettings serializerSettings)
+        public ExamplesConverter(MvcOutputFormatter mvcOutputFormatter)
         {
-            this.jsonFormatter = jsonFormatter;
             this.mvcOutputFormatter = mvcOutputFormatter;
-            this.serializerSettings = serializerSettings;
         }
 
-        public IOpenApiAny SerializeExampleXml(object exampleValue)
+        public IOpenApiAny SerializeExampleXml(object value)
         {
-            return new OpenApiString(exampleValue.XmlSerialize(mvcOutputFormatter));
+            return new OpenApiString(mvcOutputFormatter.Serialize(value, ApplicationXml).FormatXml());
         }
 
-        public IOpenApiAny SerializeExampleJson(object exampleValue)
+        public IOpenApiAny SerializeExampleJson(object value)
         {
-            return new OpenApiRawString(jsonFormatter.FormatJson(exampleValue, serializerSettings));
+            return new OpenApiRawString(mvcOutputFormatter.Serialize(value, ApplicationJson));
+        }
+
+        public IDictionary<string, OpenApiExample> ToOpenApiExamplesDictionaryXml(
+            IEnumerable<ISwaggerExample<object>> examples)
+        {
+            return ToOpenApiExamplesDictionary(examples, SerializeExampleXml);
+        }
+
+        public IDictionary<string, OpenApiExample> ToOpenApiExamplesDictionaryJson(
+            IEnumerable<ISwaggerExample<object>> examples)
+        {
+            return ToOpenApiExamplesDictionary(examples, SerializeExampleJson);
         }
 
         private static IDictionary<string, OpenApiExample> ToOpenApiExamplesDictionary(
@@ -49,18 +60,6 @@ namespace Swashbuckle.AspNetCore.Filters
                 grouping => grouping.First());
 
             return examplesDict;
-        }
-
-        public IDictionary<string, OpenApiExample> ToOpenApiExamplesDictionaryXml(
-            IEnumerable<ISwaggerExample<object>> examples)
-        {
-            return ToOpenApiExamplesDictionary(examples, SerializeExampleXml);
-        }
-
-        public IDictionary<string, OpenApiExample> ToOpenApiExamplesDictionaryJson(
-            IEnumerable<ISwaggerExample<object>> examples)
-        {
-            return ToOpenApiExamplesDictionary(examples, SerializeExampleJson);
         }
     }
 }