+++ src/Swashbuckle.AspNetCore.Filters/Examples/MvcOutputFormatter.cs
@@ -10,6 +10,7 @@ using Microsoft.Extensions.Logging;
 using Microsoft.Extensions.Options;
 using Microsoft.Net.Http.Headers;
 using System.Linq;
+using System.Text;
 
 namespace Swashbuckle.AspNetCore.Filters
 {
@@ -85,7 +86,15 @@ namespace Swashbuckle.AspNetCore.Filters
 
                 formatter.WriteAsync(outputFormatterContext).GetAwaiter().GetResult();
                 stringWriter.FlushAsync().GetAwaiter().GetResult();
-                return stringWriter.ToString();
+                var result = stringWriter.ToString();
+                if (string.IsNullOrEmpty(result))
+                {
+                    // workaround for SystemTextJsonOutputFormatter
+                    var ms = (MemoryStream)outputFormatterContext.HttpContext.Response.Body;
+                    result = Encoding.UTF8.GetString(ms.ToArray());
+                }
+
+                return result;
             }
         }