+++ src/Swashbuckle.AspNetCore.Filters/Extensions/ObjectExtensions.cs
@@ -1,55 +1,17 @@
-﻿using System;
-using System.IO;
-using System.Xml;
-using System.Xml.Linq;
-using System.Xml.Serialization;
-using Microsoft.Net.Http.Headers;
+﻿using System.Xml.Linq;
 
 namespace Swashbuckle.AspNetCore.Filters.Extensions
 {
     internal static class ObjectExtensions
     {
-        public static string XmlSerialize<T>(this T value)
+        public static string FormatXml(this string unformattedXml)
         {
-            if (value == null)
+            if (string.IsNullOrEmpty(unformattedXml))
             {
                 return string.Empty;
             }
 
-            var xmlSerializer = new XmlSerializer(value.GetType());
-            var stringWriter = new StringWriter();
-            using (var writer = XmlWriter.Create(stringWriter))
-            {
-                xmlSerializer.Serialize(writer, value);
-
-                return stringWriter
-                    .ToString()
-                    .FormatXml();
-            }
+            return XDocument.Parse(unformattedXml).ToString();
         }
-
-        private static readonly MediaTypeHeaderValue ApplicationXml = MediaTypeHeaderValue.Parse("application/xml; charset=utf-8");
-
-        internal static string XmlSerialize<T>(this T value, MvcOutputFormatter mvcOutputFormatter)
-        {
-            if (mvcOutputFormatter == null)
-            {
-                throw new ArgumentNullException(nameof(mvcOutputFormatter));
-            }
-
-            try
-            {
-                return mvcOutputFormatter
-                    .Serialize(value, ApplicationXml)
-                    .FormatXml();
-            }
-            catch (MvcOutputFormatter.FormatterNotFoundException)
-            {
-                return value.XmlSerialize();
-            }
-        }
-
-        private static string FormatXml(this string unformattedXml)
-            => XDocument.Parse(unformattedXml).ToString();
     }
 }