+++ test/Swashbuckle.AspNetCore.Filters.Test/TestFixtures/Fakes/FormatterOptions.cs
@@ -1,11 +1,30 @@
 ﻿using Microsoft.AspNetCore.Mvc;
 using Microsoft.AspNetCore.Mvc.Formatters;
 using Microsoft.Extensions.Options;
+using Newtonsoft.Json;
+using Newtonsoft.Json.Serialization;
+using System.Buffers;
+using System.Text.Json;
 
 namespace Swashbuckle.AspNetCore.Filters.Test.TestFixtures.Fakes
 {
     internal class FormatterOptions : IOptions<MvcOptions>
     {
+        private static NewtonsoftJsonOutputFormatter jsonOutputFormatter = new NewtonsoftJsonOutputFormatter(
+            new JsonSerializerSettings
+            {
+                Formatting = Formatting.Indented,
+                ContractResolver = new ExcludeObsoletePropertiesResolver(new DefaultContractResolver
+                {
+                    NamingStrategy = new CamelCaseNamingStrategy()
+                })
+            },
+            ArrayPool<char>.Shared,
+            new MvcOptions());
+
+        private static SystemTextJsonOutputFormatter systemTextJsonFormatter = new SystemTextJsonOutputFormatter(
+            new JsonSerializerOptions { WriteIndented = true });
+
         private FormatterOptions(params IOutputFormatter[] formatters)
         {
             Value = new MvcOptions();
@@ -18,7 +37,10 @@ namespace Swashbuckle.AspNetCore.Filters.Test.TestFixtures.Fakes
         public static FormatterOptions WithXmlDataContractFormatter
             => new FormatterOptions(new XmlDataContractSerializerOutputFormatter());
 
-        public static FormatterOptions WithoutFormatters
-            => new FormatterOptions();
+        public static FormatterOptions WithNewtonsoftFormatter => new FormatterOptions(jsonOutputFormatter);
+
+        public static FormatterOptions WithSystemTextJsonFormatter = new FormatterOptions(systemTextJsonFormatter);
+        public static FormatterOptions WithXmlAndNewtonsoftJsonFormatters => new FormatterOptions(new XmlSerializerOutputFormatter(), jsonOutputFormatter);
+        public static FormatterOptions WithoutFormatters => new FormatterOptions();
     }
 }